Support splitting by x and by z axis and find the split option (under flag). Solution is pretty slow (30-60 minutes to cover all tests).

+++ src/cpp/solvers/layers_base.h
@@ -42,5 +42,5 @@ protected:
 
 public:
     static uint64_t Solve(const Matrix& m, Trace& output);
-    static size_t SolveHelper(const Matrix& m, Coordinate first_and_last, Trace& output);
+    static uint64_t SolveHelper(const Matrix& m, Coordinate first_and_last, Trace& output);
 };
